NoRainbows

Documentation for NoRainbows.

Base.printstyled

NoRainbows.set_filepath

NoRainbows.set_framenumber

Base.printstyled — Method

printstyled([io], xs...; bold::Bool=false, underline::Bool=false, blink::Bool=false, reverse::Bool=false, hidden::Bool=false, color::Union{Symbol,Int}=:normal)

Print xs in a color specified as a symbol or integer, optionally in bold.

color may take any of the values :normal, :default, :bold, :black, :blink, :blue, :cyan, :green, :hidden, :light_black, :light_blue, :light_cyan, :light_green, :light_magenta, :light_red, :light_yellow, :magenta, :nothing, :red, :reverse, :underline, :white, or :yellow or an integer between 0 and 255 inclusive. Note that not all terminals support 256 colors. If the keyword bold is given as true, the result will be printed in bold. If the keyword underline is given as true, the result will be printed underlined. If the keyword blink is given as true, the result will blink. If the keyword reverse is given as true, the result will have foreground and background colors inversed. If the keyword hidden is given as true, the result will be hidden. Keywords can be given in any combination.

NoRainbows.set_filepath — Method

set_filepath(;
    default = nothing,
    filepath = nothing,
    colon = nothing,
    lineno = nothing,
    inlined = nothing
)

Set color attributes of the filepath which generate corresponding frameline.

NoRainbows.set_framenumber — Method

set_framenumber(;
    default = nothing,
    frameno = nothing,
    brackets = nothing
)

Set color attributes of the frame number where frameline was generated.